Update README.md
what I did. basically is preformed a change to the readme file, the names of the team-mates - they were incorrect, the first one was 'Name' and the second one was also 'Name'.
I corrected the mistake by replacing the text where it said 'Name' to Aran(This is the name of the first team-mate), the second appearance of the word 'Name', I changed it to 'Raz'(The second team-mate, and in fact the leader of the group).
These were all the changes I have made.
please contact 0542463025 for additional information, or email me at aranweizman@mail.huji.ac.il.
I would be happy to answer any inquiries about these changes.